pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mindroot"
version = "11.4.1"
description = "MindRoot AI Agent Framework"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "jinja2",
    "python-multipart",
    "aiofiles",
    "nanoid",
    "termcolor",
    "colored",
    "loguru",
    "pytz",
    "PyJWT",
    "pluggy",
    "sse_starlette",
    "partial-json-parser",
    "setuptools",
    "requests",
    "aiohttp",
    "Pillow",
    "psutil",
    "aiosmtplib",
    "pydantic[email]",
    "bcrypt",
    "rich",
    "colorama",
    "python-dotenv",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "chardet",
    "mcp>=1.12.3",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "uv>=0.7.0",
    "textual>=0.88.0",
    "pyperclip>=1.8.0",
    "pyinstrument"
]

[project.scripts]
mindroot = "mindroot.server:main"
mr = "mindroot.coreplugins.tui.cli:main"
